修复登录

+++ src/main/java/cn/islys/gms/repository/BackpackRepository.java
@@ -14,34 +14,34 @@ public interface BackpackRepository extends JpaRepository<Backpack, Long> {
     /**
      * 根据用户ID获取背包物品（按购买时间倒序）
      */
-    List<Backpack> findByUserIdOrderByPurchaseTimeDesc(Integer userId);
+    List<Backpack> findByUserIdOrderByPurchaseTimeDesc(Long userId);
 
     /**
      * 根据用户ID和物品ID查找背包物品
      */
-    List<Backpack> findByUserIdAndItemId(Integer userId, Long itemId);
+    List<Backpack> findByUserIdAndItemId(Long userId, Long itemId);
 
     /**
      * 根据用户ID和物品类型查找背包物品
      */
     @Query("SELECT b FROM Backpack b JOIN Item i ON b.itemId = i.id WHERE b.userId = :userId AND i.type = :type")
-    List<Backpack> findByUserIdAndItemType(@Param("userId") Integer userId, @Param("type") String type);
+    List<Backpack> findByUserIdAndItemType(@Param("userId") Long userId, @Param("type") String type);
 
     /**
      * 根据用户ID和物品名称模糊查询
      */
     @Query("SELECT b FROM Backpack b JOIN Item i ON b.itemId = i.id WHERE b.userId = :userId AND i.name LIKE %:name%")
-    List<Backpack> findByUserIdAndItemNameContaining(@Param("userId") Integer userId, @Param("name") String name);
+    List<Backpack> findByUserIdAndItemNameContaining(@Param("userId") Long userId, @Param("name") String name);
 
     /**
      * 根据用户ID、物品类型和名称查询
      */
     @Query("SELECT b FROM Backpack b JOIN Item i ON b.itemId = i.id WHERE b.userId = :userId AND i.type = :type AND i.name LIKE %:name%")
-    List<Backpack> findByUserIdAndItemTypeAndItemNameContaining(@Param("userId") Integer userId, @Param("type") String type, @Param("name") String name);
+    List<Backpack> findByUserIdAndItemTypeAndItemNameContaining(@Param("userId") Long userId, @Param("type") String type, @Param("name") String name);
 
     /**
      * 获取用户背包中存在的物品类型
      */
     @Query("SELECT DISTINCT i.type FROM Backpack b JOIN Item i ON b.itemId = i.id WHERE b.userId = :userId")
-    List<String> findDistinctItemTypesByUserId(@Param("userId") Integer userId);
+    List<String> findDistinctItemTypesByUserId(@Param("userId") Long userId);
 }
